+++ hutool-core/src/test/java/cn/hutool/core/img/ImgTest.java
@@ -20,6 +20,13 @@ public class ImgTest {
 	public void compressTest() {
 		Img.from(FileUtil.file("f:/test/4347273249269e3fb272341acc42d4e.jpg")).setQuality(0.8).write(FileUtil.file("f:/test/test_dest.jpg"));
 	}
+
+	@Test
+	@Ignore
+	public void writeTest() {
+		final Img from = Img.from(FileUtil.file("d:/test/81898311-001d6100-95eb-11ea-83c2-a14d7b1010bd.png"));
+		ImgUtil.write(from.getImg(), FileUtil.file("d:/test/dest.jpg"));
+	}
 	
 	@Test
 	@Ignore