+++ hutool-poi/src/main/java/cn/hutool/poi/excel/sax/Excel03SaxReader.java
@@ -18,6 +18,7 @@ import org.apache.poi.hssf.record.BOFRecord;
 import org.apache.poi.hssf.record.BlankRecord;
 import org.apache.poi.hssf.record.BoolErrRecord;
 import org.apache.poi.hssf.record.BoundSheetRecord;
+import org.apache.poi.hssf.record.CellValueRecordInterface;
 import org.apache.poi.hssf.record.FormulaRecord;
 import org.apache.poi.hssf.record.LabelRecord;
 import org.apache.poi.hssf.record.LabelSSTRecord;
@@ -77,7 +78,7 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 	 * 自定义需要处理的sheet编号，如果-1表示处理所有sheet
 	 */
 	private int rid = -1;
-	// 当前表索引
+	// 当前rid索引
 	private int curRid = -1;
 
 	private final RowHandler rowHandler;
@@ -193,7 +194,7 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 			if (record instanceof MissingCellDummyRecord) {
 				// 空值的操作
 				MissingCellDummyRecord mc = (MissingCellDummyRecord) record;
-				addToRowCellList(mc.getColumn(), StrUtil.EMPTY);
+				addToRowCellList(mc);
 			} else if (record instanceof LastCellOfRowDummyRecord) {
 				// 行结束
 				processLastCell((LastCellOfRowDummyRecord) record);
@@ -208,17 +209,40 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 	// ---------------------------------------------------------------------------------------------- Private method start
 
 	/**
-	 * 将单元格数据加入到行列表中
-	 * @param index 加入位置
-	 * @param value 值
+	 * 将空数据加入到行列表中
+	 *
+	 * @param record MissingCellDummyRecord
 	 */
-	private void addToRowCellList(int index, Object value){
-		while(index > this.rowCellList.size()){
+	private void addToRowCellList(MissingCellDummyRecord record) {
+		addToRowCellList(record.getRow(), record.getColumn(), StrUtil.EMPTY);
+	}
+
+	/**
+	 * 将单元格数据加入到行列表中
+	 *
+	 * @param record 单元格
+	 * @param value  值
+	 */
+	private void addToRowCellList(CellValueRecordInterface record, Object value) {
+		addToRowCellList(record.getRow(), record.getColumn(), value);
+	}
+
+	/**
+	 * 将单元格数据加入到行列表中
+	 *
+	 * @param row    行号
+	 * @param column 单元格
+	 * @param value  值
+	 */
+	private void addToRowCellList(int row, int column, Object value) {
+		while (column > this.rowCellList.size()) {
 			// 对于中间无数据的单元格补齐空白
 			this.rowCellList.add(StrUtil.EMPTY);
+			this.rowHandler.handleCell(this.curRid, row, rowCellList.size() - 1, value, null);
 		}
 
-		this.rowCellList.add(index, value);
+		this.rowCellList.add(column, value);
+		this.rowHandler.handleCell(this.curRid, row, column, value, null);
 	}
 
 	/**
@@ -232,12 +256,12 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 		switch (record.getSid()) {
 			case BlankRecord.sid:
 				// 空白记录
-				addToRowCellList(((BlankRecord) record).getColumn(), StrUtil.EMPTY);
+				addToRowCellList(((BlankRecord) record), StrUtil.EMPTY);
 				break;
 			case BoolErrRecord.sid:
 				// 布尔类型
 				final BoolErrRecord berec = (BoolErrRecord) record;
-				addToRowCellList(berec.getColumn(), berec.getBooleanValue());
+				addToRowCellList(berec, berec.getBooleanValue());
 				break;
 			case FormulaRecord.sid:
 				// 公式类型
@@ -253,7 +277,7 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 				} else {
 					value = StrUtil.wrap(HSSFFormulaParser.toFormulaString(stubWorkbook, formulaRec.getParsedExpression()), "\"");
 				}
-				addToRowCellList(formulaRec.getColumn(), value);
+				addToRowCellList(formulaRec, value);
 				break;
 			case StringRecord.sid:
 				// 单元格中公式的字符串
@@ -266,7 +290,7 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 			case LabelRecord.sid:
 				final LabelRecord lrec = (LabelRecord) record;
 				value = lrec.getValue();
-				addToRowCellList(lrec.getColumn(), value);
+				addToRowCellList(lrec, value);
 				break;
 			case LabelSSTRecord.sid:
 				// 字符串类型
@@ -274,7 +298,7 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 				if (null != sstRecord) {
 					value = sstRecord.getString(lsrec.getSSTIndex()).toString();
 				}
-				addToRowCellList(lsrec.getColumn(), ObjectUtil.defaultIfNull(value, StrUtil.EMPTY));
+				addToRowCellList(lsrec, ObjectUtil.defaultIfNull(value, StrUtil.EMPTY));
 				break;
 			case NumberRecord.sid: // 数字类型
 				final NumberRecord numrec = (NumberRecord) record;
@@ -296,7 +320,7 @@ public class Excel03SaxReader extends AbstractExcelSaxReader<Excel03SaxReader> i
 					}
 				}
 				// 向容器加入列值
-				addToRowCellList(numrec.getColumn(), value);
+				addToRowCellList(numrec, value);
 				break;
 			default:
 				break;